Revert "DO NOT REVERT: Changing to new library for robots.txt to allow me to not have to manage the nrobots myself. Lib not acting as expected https://code.google.com/p/robotstxt/issues/detail?id=6"
This reverts commit e4019052b5d27d013df2b397f75b501cae384de6.

+++ Abot/Crawler/PoliteWebCrawler.cs
@@ -41,11 +41,11 @@ namespace Abot.Crawler
 
         public override CrawlResult Crawl(Uri uri, CancellationTokenSource cancellationTokenSource)
         {
-            long robotsDotTextCrawlDelayInSecs = 0;
-            long robotsDotTextCrawlDelayInMillisecs = 0;
+            int robotsDotTextCrawlDelayInSecs = 0;
+            int robotsDotTextCrawlDelayInMillisecs = 0;
 
-            long maxRobotsDotTextCrawlDelayInSeconds = _crawlContext.CrawlConfiguration.MaxRobotsDotTextCrawlDelayInSeconds;
-            long maxRobotsDotTextCrawlDelayInMilliSecs = maxRobotsDotTextCrawlDelayInSeconds * 1000;
+            int maxRobotsDotTextCrawlDelayInSeconds = _crawlContext.CrawlConfiguration.MaxRobotsDotTextCrawlDelayInSeconds;
+            int maxRobotsDotTextCrawlDelayInMilliSecs = maxRobotsDotTextCrawlDelayInSeconds * 1000;
 
             //Load robots.txt
             if (_crawlContext.CrawlConfiguration.IsRespectRobotsDotTextEnabled)
